CP-40265 - xenopsd: Calculate max_grant_frames dynamically
Plumb the expected number of VBDs and VIFs to xenopsd's VM_create micro-op,
allowing the domain creation code to estimate how big max_grant_frames
value needs to be. Previously it was hardcoded to 64 but that was found lacking
in situations with a lot of VBDs and VIFs under load, as they'd run out of grant
table entries and the VM would crash.

Now, when fewer VIFs and VBDs are expected, fewer grant table entries are
needed, and the other way around. An expectation of 1 hotplugged VIF and VBD
each is accounted for as well, this can possibly be incremented in the future
if we decide, say, 2 hotplugged devices are common enough.

The VBD calculation is complex and approximate, as explained in the code.
Size of the ring was calculated with the following program (important because
struct sizes can change, and the code comment also outlines how this can be
improved in the future with feature detection, in which case more constants
would need to be considered), attaching it here because the BLK_RING_SIZE
macro is otherwise hard to decipher by hand:

```
// Can be compiled with just gcc test.c
// # ./test
// 64
// 112
// 32
\#include <stdint.h>
\#include <stdio.h>
\#include <stddef.h>

struct tester {
    unsigned int req_prod, req_event;
    unsigned int rsp_prod, rsp_event;
    uint8_t __pad[48];
    int ring;
};

typedef uint32_t grant_ref_t;
typedef uint16_t blkif_vdev_t;
typedef uint64_t blkif_sector_t;
\#define BLKIF_MAX_SEGMENTS_PER_REQUEST 11
\#define BLKIF_MAX_INDIRECT_PAGES_PER_REQUEST 8

struct blkif_request_segment {
		grant_ref_t gref;
		uint8_t     first_sect, last_sect;
};

struct blkif_request_rw {
	uint8_t        nr_segments;
	blkif_vdev_t   handle;
	uint32_t       _pad1;
	uint64_t       id;
	blkif_sector_t sector_number;
	struct blkif_request_segment seg[BLKIF_MAX_SEGMENTS_PER_REQUEST];
} __attribute__((__packed__));

struct blkif_request_discard {
	uint8_t        flag;
	blkif_vdev_t   _pad1;
	uint32_t       _pad2;
	uint64_t       id;
	blkif_sector_t sector_number;
	uint64_t       nr_sectors;
	uint8_t        _pad3;
} __attribute__((__packed__));

struct blkif_request_other {
	uint8_t      _pad1;
	blkif_vdev_t _pad2;
	uint32_t     _pad3;
	uint64_t     id;
} __attribute__((__packed__));

struct blkif_request_indirect {
	uint8_t        indirect_op;
	uint16_t       nr_segments;
	uint32_t       _pad1;
	uint64_t       id;
	blkif_sector_t sector_number;
	blkif_vdev_t   handle;
	uint16_t       _pad2;
	grant_ref_t    indirect_grefs[BLKIF_MAX_INDIRECT_PAGES_PER_REQUEST];
	uint32_t      _pad3;
} __attribute__((__packed__));

struct blkif_request {
	uint8_t        operation;
	union {
		struct blkif_request_rw rw;
		struct blkif_request_discard discard;
		struct blkif_request_other other;
		struct blkif_request_indirect indirect;
	} u;
} __attribute__((__packed__));

union blkif_sring_entry {                                            \
    struct blkif_request req;                                                        \
    struct blkif_request rsp;                                                        \
};

\#define __RD2(_x)  (((_x) & 0x00000002) ? 0x2                  : ((_x) & 0x1))
\#define __RD4(_x)  (((_x) & 0x0000000c) ? __RD2((_x)>>2)<<2    : __RD2(_x))
\#define __RD8(_x)  (((_x) & 0x000000f0) ? __RD4((_x)>>4)<<4    : __RD4(_x))
\#define __RD16(_x) (((_x) & 0x0000ff00) ? __RD8((_x)>>8)<<8    : __RD8(_x))
\#define __RD32(_x) (((_x) & 0xffff0000) ? __RD16((_x)>>16)<<16 : __RD16(_x))

int main () {
	size_t offset =offsetof(struct tester, ring);
	size_t size = sizeof(union blkif_sring_entry);
	printf("%d\n", offset);
	printf("%zu\n", size);
	printf("%d\n", __RD32((4096- offset)/size));
}
```

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>