[qcow-tool packaging] package it in xapi
remove dune-project from `ocaml/qcow-tool` that was here from the first
import. Now we are building qcow-tool with xapi. This patch creates
the package in the toplevel dune-project and removes uneeded build
stuff from the original repository.

To be able to build qcow-tool we added the following packages to xs-opam:
  - asetmap
  - ezjsonm
  - mirage-block-combinators
  - mirage-types (deprecated)
  - mirage-types-lwt (deprecated)
  - prometheus
  - unix-type-representations

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>