python3/qcow2-to-stdout: Update to handle block special files
The original script is designed to only work on regular files. Adapt it
slightly to work on special block devices (since that's what we want to export
from).

Aside from dropping some unnecessary checks and replacing os.path.getsize with
a functional equivalent that works on block special files, these changes remove
detection of "holes" in the raw file, since SEEK_DATA and SEEK_HOLE are
unimplemented in this case.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>