{export,import}_raw_vdi: add qcow as supported format
This patch allows to pass "qcow2" as a supported format when calling VDI
export and import.

Qcow_tool_wrapper is added as a helper that calls the Python script for export
(conversion from raw to qcow2 stream) and the OCaml Qcow_stream library for
import (conversion from qcow2 stream to raw).

vdi-import with format=qcow2 also supports not specifying the VDI (just like
format=raw). In this case, xapi will read the QCOW2 header first to determine
the virtual size and create the VDI itself before copying over the data.

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>
Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>