{export,import}_raw_vdi: add qcow as supported format
This patch allows to pass "qcow2" as a supported format when calling VDI
export and import.

Qcow_tool_wrapper is added as a helper that calls the Python script for export
(conversion from raw to qcow2 stream) and the qcow-tool CLI tool (ocaml-qcow
library) for import (conversion from qcow2 stream to raw).

Signed-off-by: Guillaume <guillaume.thouvenin@vates.tech>
Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>