xapi-idl: Make it possible for Updates.Dynamic.Vm to be more detailed
Currently, on *every* change that Xenopsd notices or does to the Vm object, an
Updates.Dynamic.Vm (uuid) is sent to xapi, which in turn triggers VM.stat,
rescanning the whole object. In testing, it was noticed that up to 41% of all
calls to xenstore during Windows 11 VM installation and boot are coming from
VM.stat.

Instead of giving a general "there has been an update to the VM object" status,
prepare for a more detailed delta to be sent, if available, to avoid
rescanning. This commit preserves the current functionality (all Dynamic.Vm
constructed have an update variant of type TotalRescan), but allows to switch
to a more detailed delta in the future.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>