xenopsd: Add a virtio-vga parameter to platform/vga
virtio-vga is a modern, VGA compatible virtio device provided by QEMU, which
the guests can be enlightened about with virtio-gpu drivers (available in
Linux since 4.14 and on Windows as one of the signed KVM drivers).

With the drivers available, this device is much faster than alternatives,
reducing latency and allows for higher guest display resolutions.

For a similar patch against upstream xen, see the one posted at
https://lore.kernel.org/xen-devel/95a54f6c4296b03db5265da50a98f03c53858677.1755088831.git.teddy.astie@vates.tech/

Adds a "virtio" option to "platform/vga", handles it in xenopsd and
qemu_wrapper. The latter required special handling since passing generic VGA
parameters to the virtio-vga device can break it, so they need to be dropped.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>