host.disable: Add host_disabled_until_reboot boolean parameter
Until now, the host_disabled_until_reboot flag in localdb was only used
internally by xapi to indicate that the host is shutting down, and that it
should not be re-enabled on xapi restarts other than the first one after a
reboot.

However, it could also be useful for manual maintenance on hosts. Currently a
manually disabled host will be re-enabled on toolstack restarts, which will
provoke VM migrations. If maintenance requires many toolstack restarts, that
could be painful.

So allow passing a boolean host_disabled_until_reboot parameter for
Host.disable, reusing the internal localdb flag for this purpose.

Also fix up the comment in xapi_host_helpers.mli.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>